--- a/sklearn/tree/setup.py
+++ b/sklearn/tree/setup.py
@@ -14,6 +14,16 @@ def configuration(parent_package="", top_path=None):
                          include_dirs=[numpy.get_include()],
                          libraries=libraries,
                          extra_compile_args=["-O3"])
+    config.add_extension("_splitter",
+                         sources=["_splitter.c"],
+                         include_dirs=[numpy.get_include()],
+                         libraries=libraries,
+                         extra_compile_args=["-O3"])
+    config.add_extension("_criterion",
+                         sources=["_criterion.c"],
+                         include_dirs=[numpy.get_include()],
+                         libraries=libraries,
+                         extra_compile_args=["-O3"])
     config.add_extension("_utils",
                          sources=["_utils.c"],
                          include_dirs=[numpy.get_include()],