This pull request handles splitting `_tree.pyx` into `_tree.pyx` with Tree and TreeBuilders, `_splitter.pyx` with all splitters and split records, `_criterion.pyx` with all criteria, and moves some functions into `_utils.pyx`. No code is changed, simply moved, in this PR. I almost have this working, with all unit tests passing except those which trigger a specific segfault. The segfault relates to the following function:

```
cdef inline np.ndarray sizet_ptr_to_ndarray(SIZE_t* data, SIZE_t size):
    """Encapsulate data into a 1D numpy array of intp's."""
    cdef np.npy_intp shape[1]
    shape[0] = <np.npy_intp> size
    return np.PyArray_SimpleNewFromData(1, shape, np.NPY_INTP, data)
```

specifically, this call:

```
np.PyArray_SimpleNewFromData(1, shape, np.NPY_INTP, data)
```

In the test case, `shape = 1`, `data = [ 2 ]`, and it triggers a segfault. I have been reading about `PyArray_SimpleNewFromData` but figured that it might be easier to ask. So, (1) do you agree with this layout, and (2) do you know why this segfault is being triggered? Thanks!

cc @glouppe @arjoly @ogrisel