@glouppe 2015-09-08
In all files using Numpy, make sure that the C-API is properly initialized using `np.import_array()`. 

See https://github.com/cython/cython/wiki/tutorials-numpy#c-api-initalization