@jmschrei 2015-09-08
Solved. You're a genius. :)