@glouppe 2015-09-08
> Solved. You're a genius. :)

I got bit by this very obscure error several times and learned from it :)