@glouppe 2015-09-09
As a double check, could you quickly make sure that there is no regression in terms of speed? (Should be all the same, but we are never too sure)