@glouppe 2015-09-09
For `_tree.pyx` the diff is too long, but it seems you removed unwillingly some comments delimiting sections, e.g. above the definition `TreeBuilder` (line 79).