@glouppe 2015-09-09
Other than my comments, I am happy with the code structure. I hope it will make our lives easier for the next PRs :)