@arjoly 2015-09-09
> (I am saying this because these changes add 12000 new lines of C code)

The culprit is cython and this is well known. Cython adds a lot of functions to handle all sort of situations without letting you know.