@jmschrei 2015-09-09
All changes incorporated, and all unit tests pass on my computer. Lets see if travis/appveyor build.