@arjoly 2015-09-09
A last nitpick if you agree, can you sort import alphabetically?

Thanks a lot for this pr.