@jmschrei 2015-09-09
Times seem to be the same as well. All classifiers use default settings and `n_estimators=10` if they are an ensemble.

I'll alphabetize imports now.

```
BRANCH
RandomForestClassifier
madelon    0.623 0.26
Gaussian   0.914 4.072
arcene     0.76 0.02
spambase   0.941 0.046
mnist      0.948 4.902
covtypes   0.945 14.429

ExtraTreesClassifier
madelon    0.592 0.079
Gaussian   0.919 0.657
arcene     0.73 0.013
spambase   0.945 0.041
mnist      0.95 4.642
covtypes   0.943 12.037

GradientBoostingClassifier
madelon    0.74 0.657
Gaussian   0.853 8.004
arcene     0.74 0.843
spambase   0.911 0.092
mnist      0.836 165.77
covtypes   0.706 95.89

DecisionTreeClassifier
madelon    0.74 0.634
Gaussian   0.734 10.386
arcene     0.63 0.192
spambase   0.917 0.065
mnist      0.882 22.677
covtypes   0.943 10.317

GradientBoostingRegressor
boston 0.0 0.008
regression 0.0 1.778
diabetes 0.0 0.006

ExtraTreesRegressor
boston 0.02 0.019
regression 0.0 3.068
diabetes 0.0 0.015

RandomForestRegressor
boston 0.0 0.03
regression 0.0 10.94
diabetes 0.0 0.022

DecisionTreeRegressor
boston 0.02 0.004
regression 0.0 1.723
diabetes 0.0 0.003

MASTER
RandomForestClassifier
madelon    0.623 0.262
Gaussian   0.914 4.53
arcene     0.76 0.029
spambase   0.941 0.051
mnist      0.948 5.41
covtypes   0.945 14.127

ExtraTreesClassifier
madelon    0.592 0.088
Gaussian   0.919 0.75
arcene     0.73 0.014
spambase   0.945 0.045
mnist      0.95 4.926
covtypes   0.943 12.366

GradientBoostingClassifier
madelon    0.74 0.667
Gaussian   0.853 8.049
arcene     0.74 0.858
spambase   0.911 0.092
mnist      0.836 168.6
covtypes   0.706 97.512

DecisionTreeClassifier
madelon    0.74 0.637
Gaussian   0.734 10.428
arcene     0.63 0.191
spambase   0.917 0.066
mnist      0.882 22.735
covtypes   0.943 10.363

GradientBoostingRegressor
boston 0.0 0.007
regression 0.0 1.691
diabetes 0.0 0.006

ExtraTreesRegressor
boston 0.02 0.017
regression 0.0 2.888
diabetes 0.0 0.014

RandomForestRegressor
boston 0.0 0.03
regression 0.0 10.875
diabetes 0.0 0.022

DecisionTreeRegressor
boston 0.02 0.004
regression 0.0 1.762
diabetes 0.0 0.003
```