@arjoly 2015-09-09
Could you run `benchmarks/bench_covertype.py`?