@jmschrei 2015-09-09
Windows is not happy. I think maybe I'm not allowed to import inline functions?