@jmschrei 2015-09-09
Everything looks good, commits have been squashed. @arjoly do you give your +1?