@jmschrei 2015-09-09
```
benchmark/bench_covertype

Classification performance:
===========================
Classifier   train-time test-time error-rate
--------------------------------------------
CART          12.0291s   0.0214s     0.0424  
liblinear     8.8294s    0.0124s     0.2305  
SGD           0.5443s    0.0170s     0.2315  
GaussianNB    0.2599s    0.0788s     0.4841 
```