@arjoly 2015-09-09
I was thinking to this benchmark.

```
± make in &&  python benchmarks/bench_covertype.py --classifiers CART ExtraTrees RandomForest GBRT

# Master

Classifier   train-time test-time error-rate
--------------------------------------------
RandomForest  45.6812s   0.3655s     0.0330  
ExtraTrees    34.7320s   0.5940s     0.0372  
CART          11.7957s   0.0919s     0.0424  
GBRT         446.7187s   0.2867s     0.1777  

# this pr

Classifier   train-time test-time error-rate
--------------------------------------------
RandomForest  47.4927s   0.3775s     0.0330  
ExtraTrees    33.4654s   0.5784s     0.0372  
CART          12.3618s   0.0693s     0.0424  
GBRT         447.3703s   0.2879s     0.1777  
```

Overall it looks good. I am merging.