@arjoly 2015-09-09
thanks @jmschrei I am eager to see new speed up! :-)